Revert "109 add checks to code (#124)" (#129)
This reverts commit a2131e9c531daf59fe872e7e227a10279a8605c2.

+++ server/api/application/db/DB.php
@@ -196,13 +196,9 @@ class DB
 
     
 
-    public function deleteUnit($unitId, $villageId)
+    public function deleteUnit($unitId, $userId)
     {
-        return $this->execute("
-            DELETE FROM units 
-            WHERE id = ? AND village_id = ?", 
-            [$unitId, $villageId]
-        );
+        return $this->execute("DELETE FROM units WHERE id = ? AND user_id = ?", [$unitId, $userId]);
     }
 
     public function getBuildings($userId)
@@ -229,15 +225,6 @@ class DB
         $this->execute("INSERT INTO buildings
             (type_id, village_id, x, y, current_hp) VALUES (?, ?, ?, ?, ?)", 
             [$buildingId, $villageId, $x, $y, $hp]
-        );
-    }
-
-    public function getBuilding($buildingId, $villageId) {
-        return $this->query(
-            "SELECT id 
-            FROM buildings
-            WHERE id = ? AND village_id = ?",
-            [$buildingId, $villageId]
         );
     }
 
@@ -273,12 +260,14 @@ class DB
         );
     }
 
-    public function deleteBuilding($buildingId, $villageId)
+    public function deleteBuilding($buildingId, $userId)
     {
         return $this->execute(
-            "DELETE FROM buildings
-            WHERE id = ? AND village_id = ?",
-            [$buildingId, $villageId]
+            "
+            DELETE FROM buildings 
+            WHERE id = ? 
+            AND village_id = (SELECT id FROM villages WHERE user_id = ?)",
+            [$buildingId, $userId]
         );
     }
